Create RevisionDetector class.
- Move revision implementation out of SlicingEngine.java.
